requirements/chatbot.txt
# Dependências para o Chatbot ANTAQ
# Core ML e AI
openai>=1.12.0
tiktoken>=0.5.0

# Interface e Web
streamlit>=1.31.0
streamlit-chat>=0.1.1

# Banco vetorial e busca
chromadb>=0.4.22
sentence-transformers>=2.2.2

# Processamento de dados
pandas>=2.1.0
numpy>=1.24.3
pyarrow>=14.0.0

# Processamento de texto
langchain>=0.1.0
langchain-openai>=0.0.8
langchain-community>=0.0.19

# Utilitários
python-dotenv>=1.0.0
tqdm>=4.66.0
dateparser>=1.2.0

# Interface aprimorada
plotly>=5.17.0
altair>=5.2.0

# Cache e performance
diskcache>=5.6.3

# Logging e monitoramento
loguru>=0.7.2